bleh

+++ app/src/main/java/com/example/cookme/addItem.java
@@ -111,11 +111,11 @@ public class addItem extends AppCompatActivity {
 
                                 if(key!=-1)
                                 {
-                                    long currAmount = ingredientList.get(key).getAmount();
-                                    long newAmount = currAmount + LAmount;
+                                    double currAmount = ingredientList.get(key).getAmount();
+                                    double newAmount = currAmount + LAmount;
                                     //update data syntax:
                                     //mDatabase.child("users").child(userId).child("username").setValue(name);
-                                    ingredientList.get(key).setAmount(newAmount);
+                                    ingredientList.get(key).setAmount((Long)newAmount);
 
                                     mRootRef.child("Inventory").child(groupId).setValue(ingredientList);
                                 }
